% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{add_subset}
\alias{add_subset}
\title{Adds subset column to student activity dataframes}
\usage{
add_subset(student_df)
}
\arguments{
\item{student_df}{A dataframe containing a student's map and khan activity}
}
\value{
returns dataframe with subset column
}
\description{
Adds subset column to student activity dataframes
}

