% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_khanmath}
\alias{clean_khanmath}
\title{Clean Khan exercise data so it can be joined with state change and MAP data.}
\usage{
clean_khanmath(khanmath, numeric_sid = TRUE)
}
\arguments{
\item{khanmath}{dataframe containing composite exercise data for students.
should look like dataframe in data(simulated_khanmath)}

\item{numeric_sid}{conditional indicating whether studentid is numeric or
not. default is TRUE}
}
\value{
returns a cleaned dataframe
}
\description{
Clean Khan exercise data so it can be joined with state change and MAP data.
}

