% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_map}
\alias{clean_map}
\title{Clean MAP data so it can be joined with Khan exercise and state change data.}
\usage{
clean_map(mapdata, numeric_sid = TRUE)
}
\arguments{
\item{numeric_sid}{conditional indicating whether studentid is numeric or
not. default is TRUE}

\item{map}{Dataframe of MAP data provided by NWEA. Looks like
dataframe in data(sample_map_data)}
}
\value{
returns a cleaned dataframe
}
\description{
Clean MAP data so it can be joined with Khan exercise and state change data.
}

