% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_states}
\alias{clean_states}
\title{Cleans Khan state change data so it can be joined with exercise and map data.}
\usage{
clean_states(khanstates, numeric_sid = TRUE)
}
\arguments{
\item{khanstates}{dataframe containing state change data for students. should
look like dataframe in data(simulated_khanstate.rda)}

\item{numeric_sid}{conditional indicating whether studentid is numeric or
not. default is TRUE}
}
\value{
returns a cleaned dataframe
}
\description{
Cleans Khan state change data so it can be joined with exercise and map data.
}

