% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_activity.R
\name{generate_activity}
\alias{generate_activity}
\title{Clean and merge MAP, Khan, and item difficulty data.}
\usage{
generate_activity(map, khan_exers, khan_states, numeric_sid = TRUE,
  estimate_type = "tree", inhouse_est = NA, verbose = TRUE)
}
\arguments{
\item{map}{Dataframe containing student MAP test data provided by NWEA. Looks
like dataframe found in data(sample_map_data)}

\item{khan_exers}{Dataframe containing student data for Khan Academy excises.
Looks like dataframe found in data(simulated_khanmath)}

\item{khan_states}{Dataframe containing student state change data for Khan
Academy exercises. Looks like dataframe found in data(simulated_khanstate)}

\item{numeric_sid}{Conditional indicating whether studentid is numeric or
not. Default is TRUE}

\item{estimate_type}{String to indicate which type (NWEA, tree, inhouse)
of Khan item estimate to use. Defaults to tree (more coverage than nwea). If
'inhouse' is chosen, user must provide filepath to csv of inhouse estimates
containing exercise title and estimated difficulty. Can be created using
update_diff() function.}

\item{verbose}{Default is TRUE to print status updates.}
}
\value{
Returns list of dataframes. Each dataframe will contain a student's
Khan Activity as well as his / her MAP test history.
}
\description{
Clean and merge MAP, Khan, and item difficulty data.
}

