% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{get_item_history}
\alias{get_item_history}
\title{Get an item's mastery history}
\usage{
get_item_history(student_activity, item_title)
}
\arguments{
\item{student_activity}{A list of dataframes created by generate_activty()}

\item{item_title}{name of Khan exercise to get history for}
}
\value{
returns a dataframe containing mastery history and rit estimate
}
\description{
Get an item's mastery history
}

