% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_wide_states}
\alias{parse_wide_states}
\title{Take Khan state change data in wide format and turn it into student
activty dataframe.}
\usage{
parse_wide_states(states_wide)
}
\arguments{
\item{states_wide}{Dataframe of Khan state change data in wide format
generated by reshape_states()}
}
\value{
a parsed dataframe
}
\description{
Take Khan state change data in wide format and turn it into student
activty dataframe.
}

