% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_activity.R
\name{plot_activity}
\alias{plot_activity}
\title{Creates plot showing MAP test scores and khan activity.}
\usage{
plot_activity(stu_activity_df)
}
\arguments{
\item{stu_activity_df}{a dataframe created by generate_activity()}
}
\value{
prints a plot
}
\description{
Creates plot showing MAP test scores and khan activity.
}

