% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{populate_map}
\alias{populate_map}
\title{Add needed columns to student's map data to concatenate with state change
data}
\usage{
populate_map(student_map)
}
\arguments{
\item{student_map}{dataframe containing studentid, teststartdate, and
testritscore}
}
\value{
returns list of dataframes
}
\description{
Add needed columns to student's map data to concatenate with state change
data
}

