% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{reshape_states}
\alias{reshape_states}
\title{Reshapes Khan state data that has been cleaned into wide format to generate
Khan Activity for each student}
\usage{
reshape_states(states_long)
}
\arguments{
\item{states_long}{dataframe containing cleaned Khan state change data}
}
\value{
returns a reshaped dataframe
}
\description{
Reshapes Khan state data that has been cleaned into wide format to generate
Khan Activity for each student
}

