% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_diffs.R
\name{update_diffs}
\alias{update_diffs}
\title{Uses student activity data to update item difficulties.}
\usage{
update_diffs(student_activity, priors = "tree", priorpath = NA,
  filepath = NA, numeric_sid = TRUE, plot = FALSE, verbose = TRUE)
}
\arguments{
\item{student_activity}{List of dataframes returned by generate_activity()}

\item{priors}{A string (nwea, tree, inhouse), to indicate which estimates to
use as priors. Defaults to tree.}

\item{priorpath}{If inhouse priors are used, provide path to csv containing
these estimates.}

\item{filepath}{a path and filename to write the updated difficulties to.
ex: '/mypath/my_estimates.csv'}

\item{numeric_sid}{Conditional indicating whether studentid is numeric or
not. Default is TRUE}

\item{plot}{TRUE if you want to print a plot showing change in difficulty.
Default is FALSE.}

\item{verbose}{Default is TRUE to print status updates.}
}
\value{
returns dataframe of updated difficulties or writes to a csv if
filepath is not NA
}
\description{
Uses student activity data to update item difficulties.
}

