% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_proficiencies.R
\name{update_proficiencies}
\alias{update_proficiencies}
\title{Uses the ELO Rating system to estimate a student's current proficiecy level
on the RIT scale using his / her Khan activity.}
\usage{
update_proficiencies(studentids, student_activity, start_date, end_date = NA)
}
\arguments{
\item{studentids}{the studentids of the student proficiencies you want
to estimate}

\item{student_activity}{list of dataframes returned by generate_activity()}

\item{start_date}{date of MAP test used as initial estimate}

\item{end_date}{date to project until. Not required unless start_date is
before most recent MAP test event}
}
\value{
returns a dataframe containing studentids and estimated proficiency
}
\description{
Uses the ELO Rating system to estimate a student's current proficiecy level
on the RIT scale using his / her Khan activity.
}

